CRUD Batch

+++ database/migrations/2026_05_18_082236_create_vocab_progress_table.php
@@ -7,18 +7,27 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration
 {
     /**
+     * Determine if the migration should be run within a transaction.
+     *
+     * @var bool
+     */
+    public $withinTransaction = false;
+
+    /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::create('vocab_progress', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('id_user')->constrained('users')->onDelete('cascade');
-            $table->string('vocabulary_id');
-            $table->boolean('is_memorized')->default(false);
-            $table->timestamps();
-            $table->unique(['id_user', 'vocabulary_id']);
-        });
+        if (!Schema::hasTable('vocab_progress')) {
+            Schema::create('vocab_progress', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('id_user')->constrained('users')->onDelete('cascade');
+                $table->string('vocabulary_id');
+                $table->boolean('is_memorized')->default(false);
+                $table->timestamps();
+                $table->unique(['id_user', 'vocabulary_id']);
+            });
+        }
     }
 
     /**
